+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Scheduled event helper to ping honeybadger after task is complete
+
 ### Changes
 * Changes exception type reported by the test command for more clarity in the HB application UI.